testing/requirements.txt
code dependencies: 

anytree - https://pypi.org/project/anytree/
numpy 
pytorch
pandas 
